+++ observation_portal/requestgroups/test/test.py
@@ -1,17 +1,22 @@
 from django.utils import timezone
 from django.test import TestCase
 from mixer.backend.django import mixer
+from rest_framework.serializers import ValidationError
 from datetime import datetime
+from unittest.mock import patch
 import math
+import copy
 
 from observation_portal.requestgroups.models import (
     Request, Configuration, Target, RequestGroup, Window, Location, Constraints, InstrumentConfig,
     AcquisitionConfig, GuidingConfig
 )
 from observation_portal.proposals.models import Proposal, TimeAllocation, Semester
-from observation_portal.common.configdb import ConfigDBException
+from observation_portal.common.configdb import ConfigDBException, configdb
 from observation_portal.common.test_helpers import SetTimeMixin
 from observation_portal.requestgroups.duration_utils import PER_CONFIGURATION_STARTUP_TIME, PER_CONFIGURATION_GAP
+from observation_portal.requestgroups.serializers import InstrumentTypeValidationHelper, ModeValidationHelper
+from observation_portal.requestgroups.test.test_api import generic_payload
 
 
 class TestRequestGroupTotalDuration(SetTimeMixin, TestCase):
@@ -452,3 +457,105 @@ class TestRequestDuration(SetTimeMixin, TestCase):
         with self.assertRaises(ConfigDBException) as context:
             _ = self.configuration_expose.duration
             self.assertTrue('not found in configdb' in context.exception)
+
+
+class TestValidationHelper(TestCase):
+    def setUp(self) -> None:
+        self.mock_instrument_type = {'code': '1M0-SCICAM-SBIG',
+                                     'validation_schema': {'extra_params': {'type': 'dict',
+                                                                            'schema': {'defocus': {'type': 'float',
+                                                                                                   'min': -5.0,
+                                                                                                   'max': 5.0}}},
+                                                           'exposure_time': {'type': 'integer', 'min': 0}}}
+
+        self.generic_payload = copy.deepcopy(generic_payload)
+        self.request_instrument_type = self.generic_payload['requests'][0]['configurations'][0]['instrument_type']
+        self.instrument_config = self.generic_payload['requests'][0]['configurations'][0]['instrument_configs'][0]
+        self.guiding_config = self.generic_payload['requests'][0]['configurations'][0]['guiding_config']
+        self.acquisition_config = self.generic_payload['requests'][0]['configurations'][0]['acquisition_config']
+        self.muscat_extra_params = {'exposure_time_g': 60,
+                                    'exposure_time_r': 90,
+                                    'exposure_time_i': 60,
+                                    'exposure_time_z': 120,
+                                    'exposure_mode': 'SYNCHRONOUS'}
+
+    @patch('observation_portal.requestgroups.serializers.configdb.get_instrument_type_by_code')
+    def test_validate_instrument_config_and_extra_params_good_config(self, mock_instrument_type):
+        instrument_config = self.instrument_config.copy()
+        mock_instrument_type.return_value = self.mock_instrument_type
+        instrument_config['extra_params'] = {'defocus': 2.0}
+
+        validation_helper = InstrumentTypeValidationHelper(self.request_instrument_type)
+        validated_config = validation_helper.validate(instrument_config)
+
+        self.assertEqual(instrument_config, validated_config)
+
+    @patch('observation_portal.requestgroups.serializers.configdb.get_instrument_type_by_code')
+    def test_validate_instrument_config_and_extra_params_bad_config(self, mock_instrument_type):
+        instrument_config = self.instrument_config.copy()
+        mock_instrument_type.return_value = self.mock_instrument_type
+        instrument_config['extra_params'] = {'defocus': 2.0}
+        instrument_config['exposure_time'] = -20
+
+        validation_helper = InstrumentTypeValidationHelper(self.request_instrument_type)
+        with self.assertRaises(ValidationError) as e:
+            validation_helper.validate(instrument_config)
+        self.assertIn('exposure_time', str(e.exception))
+
+    def test_validate_mode_config_filled_in_when_missing(self):
+        guiding_config = {}
+        modes = configdb.get_modes_by_type(self.request_instrument_type)
+        validation_helper = ModeValidationHelper('guiding', self.request_instrument_type, modes['guiding'])
+
+        validated_config = validation_helper.validate(guiding_config)
+        self.assertEqual(validated_config['mode'], 'OFF')
+
+    def test_validate_mode_config_good_config(self):
+        instrument_config = self.instrument_config.copy()
+        instrument_config['mode'] = "1m0_sbig_1"
+        modes = configdb.get_modes_by_type(self.request_instrument_type)
+
+        validation_helper = ModeValidationHelper('readout', self.request_instrument_type, modes['readout'])
+        validated_config = validation_helper.validate(instrument_config)
+
+        self.assertEqual(validated_config, instrument_config)
+
+    def test_validate_mode_config_bad_config(self):
+        instrument_config = self.instrument_config.copy()
+        instrument_config['mode'] = "1m0_sbig_2"
+        instrument_config['bin_x'] = 1
+        instrument_config['bin_y'] = 2
+        modes = configdb.get_modes_by_type(self.request_instrument_type)
+
+        validation_helper = ModeValidationHelper('readout', self.request_instrument_type, modes['readout'])
+        with self.assertRaises(ValidationError) as e:
+            validation_helper.validate(instrument_config)
+        self.assertIn('bin_x', str(e.exception))
+
+    def test_validate_extra_param_mode_good_config(self):
+        instrument_config = self.instrument_config.copy()
+        instrument_type = "2M0-SCICAM-MUSCAT"
+        instrument_config['instrument_type'] = instrument_type
+        instrument_config['extra_params'] = self.muscat_extra_params
+        modes = configdb.get_modes_by_type(instrument_type)
+
+        validation_helper = ModeValidationHelper('exposure', instrument_type, modes['exposure'],
+                                                 is_extra_param_mode=True)
+        validated_instrument_config = validation_helper.validate(instrument_config)
+
+        self.assertEqual(instrument_config, validated_instrument_config)
+
+    def test_validate_extra_param_mode_bad_config(self):
+        instrument_config = self.instrument_config.copy()
+        instrument_type = "2M0-SCICAM-MUSCAT"
+        instrument_config['instrument_type'] = instrument_type
+        instrument_config['extra_params'] = self.muscat_extra_params.copy()
+        del instrument_config['extra_params']['exposure_mode']
+        modes = configdb.get_modes_by_type(instrument_type)
+
+        validation_helper = ModeValidationHelper('exposure', instrument_type, modes['exposure'],
+                                                 is_extra_param_mode=True)
+
+        with self.assertRaises(ValidationError) as e:
+            validation_helper.validate(instrument_config)
+        self.assertIn('exposure_mode', str(e.exception))
